[LIVY-678] Thrift ldap authentication, based on ldapurl, basedn, domain
## What changes were proposed in this pull request?
Currently, livy thrift-server doesn't support LDAP Authentication. We need to add LDAP authentication as that's preferable method due to security reasons.
The authentication was based on ldapUrl, basedn, domain.
User-based filters are also supported in this PR.

## How was this patch tested?

UTs tests for this part have been added. We can test in UTs

Author: captainzmc <micahzhao@tencent.com>

Closes #236 from captainzmc/thrift-ldap-impl.